minor changes.
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@193 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/quadrature/test/quad_test.cc
@@ -1,5 +1,6 @@
-#include "quad_test.hh"
+#include<config.h>
 
+#include "quad_test.hh"
 #include <set>
 
 namespace Dune {
@@ -251,7 +252,7 @@ namespace Dune {
 
   void Quad_Test::indicesTest() 
   {
-    int id;
+    size_t id;
     {
       Quadrature<double, 1> quadTemp(line, 5);
       id = quadTemp.id();
